requirement.txt
streamlit
authlib
python-dotenv
pymongo
lanchain
langchain-huggingface
langchain-pinecone